+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Use standard FOLIO UX for Delete Course (in action menu). Fixes UICR-201.
 * When deleting a course that has no cross-listed courses, also delete its underlying courselisting. This ensures orphan courselistings are not created. Fixes UICR-206.
 * With orphan courselistings no longer remaining, it is now always possible to delete terms. Fixes UICR-203.
+* App Context menu | Add new option | Courses app search. Refs UICR-166.
 
 ## [6.0.3](https://github.com/folio-org/ui-courses/tree/v6.0.3) (2023-10-20)